requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
Pillow==10.0.0
python-docx==0.8.11
Werkzeug==2.3.7
pdf2docx==0.5.6
reportlab==4.0.4
PyMuPDF==1.23.26